pyproject.toml
[tool.poetry]
name = "degradai"
version = "0.1.0"
description = ""
authors = ["patacoing <romain.lefebvre1104@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10.15"
pytest = "^8.3.3"
scikit-learn = "^1.5.2"
ecotag-sdk = "^0.1.2"
azure-ai-ml = "^1.22.1"
azure-identity = "^1.19.0"
mlflow = "^2.17.2"
pillow = "^11.0.0"
opencv-python-headless = "^4.10.0.84"
tensorflow = "2.16.2"
keras = "^3.6.0"
fastapi = {extras = ["standard"], version = "^0.115.5"}
python-multipart = "^0.0.17"
pydantic = "^2.9.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
tensorflow-metal = "^1.1.0"
tensorflow-macos = "^2.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
